Got the timer working!
textmesh pro wasnt working with the visual script nodes, so I went to the project settings, then visual scripting, then added "Text" to the type options. Then regenerated Nodes. Now the timer shows up on the screen. And, when the timer runs out, it checks to see if the player has the knife, and if he doesnt, it gives it to him. But if he doesnt, it takes it away.